Erfan: Check the output model of the GSI to add in message
If the Model is empty, that method will send "Generic", if "qssi" of Qualcomm Generic, will be "QSSI (Qualcomm Generic)". If isn't empty or QSSI, may will be the model or "Generic", depends if the text output is correct or no.
The proof of test: https://t.me/yumigsis_group/45001

Signed-off-by: Enzo Vieira | YukoSky <daffetyxd@gmail.com>